parse_config: add a new 'operator' field in s_p_values_t
Provide the capability to manage key[+-*/]=val entries in the parsing logic.
A new operator field now enables to know which operation is associated to
a pair of key/value instead of the classical set. Introduced operators are
provided by the slurm_parser_operator_t type with the following value :

 * S_P_OPERATOR_ADD
 * S_P_OPERATOR_SUB
 * S_P_OPERATOR_MUL
 * S_P_OPERATOR_DIV

It is the caller of the parsing logic that is responsible of using that
information in order to act as requested. The parsing logic do not do any
computation based on the extracted operator.

Why:

Introducing the layouts framework, the parsing logic of SLURM was extended to
manage efficiently conf elements expansions. Layouts heavily rely on the SLURM
parser to tansform textual configurations into internal representations.

Looking at a way to provide live update of Layouts information, it was
interesting to reuse the parsing logic of SLURM instead of trying to do the
same thing somewhere else.

Introducing a "key[+-*/]=val" feature was thus mandatory in order to apply
numerical modifications of the layouts internal states on the controller side
instead of trying to compute that on the client side and then send the updated
information, loosing the capability to perform operations atomicallly.