parser: add s_p_parse_buffer to parse configuration lines from a Slurm Buffer
s_p_parse_file is a convenient way to handle the parsing of Slurm configuration
file. It however only works with traditional file and can not be used to process
a bunch of configuration strings.

s_p_parse_buffer provides the capability to parse a bunch of contiguous strings
packed into a Buf type, stopping the processing at the first unsuccessfull
string extraction from the buffer.