layouts: extend API with layouts_entity_get_mkv[_ref] calls
The new calls enable to perform the get operations on a set of key/value pairs of
a particular entity in a defined layout. (mkv stands for multiple k/v pairs)

When _ref is suffixed, then the calls only deal with references and no copy of
the data is performed, only data references (pointers) are got. The _ref suffixed
calls should then be considered as more efficient but can lead to issues if the
pointers are modified after the get operations occur but before their real use.

For that reason, layouts_{lock,unlock} calls would certainly appear in a next
commit as well as _nolock suffixed calls in order to better handle these
scenarios.