hostlist: set MAX_PREFIX_CNT to 64k instead of 1k
Having MAX_PREFIX_CNT to 1k limits the number of elements that can be
expressed using expressions like node[1-1025].ib[0] to 1024.

As 64k is the max number of nodes that can be handled by Slurm, increase
that value to 64k in order to be able to express at least that number of
elements using a[x]b[y] when y is the same value for all the elements.

Note: also required by the layouts parsing logic when parsing float values
in expressions like : 'Entity=node[1-64000] FloatKey=[1-64000].[0]'